packages/error-contracts/pyproject.toml
[project]
name = "error-contracts"
version = "0.1.0"
description = "Error code definitions, codegen, and translation validation"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
